+++ extensions/mailersend/src/domains.tsx
@@ -2,7 +2,7 @@ import { Action, ActionPanel, Color, Icon, List } from "@raycast/api";
 import { getFavicon } from "@raycast/utils";
 import { useMailerSendPaginated } from "./mailersend";
 import { Activity, ActivityEventType, Domain } from "./interfaces";
-import { useMemo } from "react";
+import { useMemo, useState } from "react";
 import dayjs from "dayjs";
 import localizedFormat from "dayjs/plugin/localizedFormat";
 dayjs.extend(localizedFormat);
@@ -35,6 +35,7 @@ export default function Domains() {
 }
 
 function Activities({ domain }: { domain: Domain }) {
+  const [filter, setFilter] = useState("");
   const from = useMemo(() => dayjs().subtract(1, "day").unix(), []);
   const to = useMemo(() => dayjs().unix(), []);
 
@@ -42,26 +43,36 @@ function Activities({ domain }: { domain: Domain }) {
     [ActivityEventType.OPENED]: Color.Green,
     [ActivityEventType.DELIVERED]: Color.Blue,
     [ActivityEventType.QUEUED]: undefined,
-    [ActivityEventType.SENT]: undefined,
-    [ActivityEventType.SOFT_BOUNCED]: undefined,
+    [ActivityEventType.SENT]: Color.Purple,
+    [ActivityEventType.SOFT_BOUNCED]: Color.Orange,
     [ActivityEventType.HARD_BOUNCED]: Color.Red,
     [ActivityEventType.CLICKED]: undefined,
     [ActivityEventType.UNSUBSCRIBED]: undefined,
     [ActivityEventType.SPAM_COMPLAINTS]: undefined,
   };
 
+  
   const {
     isLoading,
     data: activities,
     pagination,
   } = useMailerSendPaginated<Activity>(`activity/${domain.id}?date_from=${from}&date_to=${to}`);
+  
+  const filteredActivities = !filter ? activities : activities.filter(a => a.type===filter);
+  
   return (
-    <List isLoading={isLoading} pagination={pagination}>
+    <List isLoading={isLoading} pagination={pagination} searchBarAccessory={<List.Dropdown tooltip="Event status" onChange={setFilter}>
+      <List.Dropdown.Item icon={Icon.Dot} title="All" value="" />
+      <List.Dropdown.Section>
+      {Object.entries(ActivityEventType).map(([key, val]) => <List.Dropdown.Item key={key} icon={{ source: Icon.Dot, tintColor: TYPE_TO_COLOR[val] }} title={key} value={val} />)}
+      </List.Dropdown.Section>
+    </List.Dropdown>}>
+      {!isLoading && !filteredActivities.length && <List.EmptyView icon={Icon.MagnifyingGlass} title="No activity found" description="Please try again with other keywords, filters or set a different period." />}
       <List.Section
         title={domain.name}
         subtitle={`${dayjs.unix(from).format("lll")} - ${dayjs.unix(to).format("lll")}`}
       >
-        {activities.map((activity) => (
+        {filteredActivities.map((activity) => (
           <List.Item
             key={activity.id}
             icon={{ value: { source: Icon.Dot, tintColor: TYPE_TO_COLOR[activity.type] }, tooltip: activity.type }}