+++ extensions/mailersend/CHANGELOG.md
@@ -1,5 +1,12 @@
 # MailerSend Changelog
 
+## [Manage Tokens + Filter Activity] - {PR_MERGE_DATE}
+
+- Filter Domain Activity by Type
+- View Tokens
+    - Rename Token
+    - Toggle (Pause/Unpause) Token
+
 ## [Initial Version] - 2025-06-25
 
 - View Domains